+++ test/cmdlineTests/yul_optimizer_steps_with_empty_cleanup_sequence/input.sol
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.0;
+pragma abicoder v2;
+
+contract C {
+    constructor() payable {
+        assembly ("memory-safe") {
+            let a := 0
+            // Without the cleanup sequence this will not be simplified to ``revert(a, a)``.
+            revert(0, a)
+        }
+    }
+}