+++ Changelog.md
@@ -4,6 +4,7 @@ Language Features:
 
 
 Compiler Features:
+* Yul Optimizer: Allow replacing the previously hard-coded cleanup sequence by specifying custom steps after a colon delimiter (``:``) in the sequence string.
 
 
 Bugfixes: